% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{create_resolver_endpoint}
\alias{create_resolver_endpoint}
\title{Creates a resolver endpoint}
\usage{
create_resolver_endpoint(
  CreatorRequestId = "string",
  Name = "string",
  SecurityGroupIds = list(
    "string"
  ),
  Direction = "INBOUND"|"OUTBOUND",
  IpAddresses = list(
    list(
      SubnetId = "string",
      Ip = "string"
    )
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{CreatorRequestId}{[required] A unique string that identifies the request and that allows failed requests to be retried without the risk of executing the operation twice. \code{CreatorRequestId} can be any unique string, for example, a date/time stamp.}

\item{Name}{A friendly name that lets you easily find a configuration in the Resolver dashboard in the Route 53 console.}

\item{SecurityGroupIds}{[required] The ID of one or more security groups that you want to use to control access to this VPC. The security group that you specify must include one or more inbound rules (for inbound resolver endpoints) or outbound rules (for outbound resolver endpoints).}

\item{Direction}{[required] Specify the applicable value:
\itemize{
\item \code{INBOUND}: Resolver forwards DNS queries to the DNS service for a VPC from your network or another VPC
\item \code{OUTBOUND}: Resolver forwards DNS queries from the DNS service for a VPC to your network or another VPC
}}

\item{IpAddresses}{[required] The subnets and IP addresses in your VPC that you want DNS queries to pass through on the way from your VPCs to your network (for outbound endpoints) or on the way from your network to your VPCs (for inbound resolver endpoints).}

\item{Tags}{A list of the tag keys and values that you want to associate with the endpoint.}
}
\description{
Creates a resolver endpoint. There are two types of resolver endpoints, inbound and outbound:
}
\details{
\itemize{
\item An \emph{inbound resolver endpoint} forwards DNS queries to the DNS service for a VPC from your network or another VPC.
\item An \emph{outbound resolver endpoint} forwards DNS queries from the DNS service for a VPC to your network or another VPC.
}
}
\examples{

}
