% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{delete_resolver_endpoint}
\alias{delete_resolver_endpoint}
\title{Deletes a resolver endpoint}
\usage{
delete_resolver_endpoint(ResolverEndpointId)
}
\arguments{
\item{ResolverEndpointId}{[required] The ID of the resolver endpoint that you want to delete.}
}
\description{
Deletes a resolver endpoint. The effect of deleting a resolver endpoint depends on whether it's an inbound or an outbound resolver endpoint:
}
\details{
\itemize{
\item \strong{Inbound}: DNS queries from your network or another VPC are no longer routed to the DNS service for the specified VPC.
\item \strong{Outbound}: DNS queries from a VPC are no longer routed to your network or to another VPC.
}
}
\section{Accepted Parameters}{
\preformatted{delete_resolver_endpoint(
  ResolverEndpointId = "string"
)
}
}

