% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{delete_resolver_rule}
\alias{delete_resolver_rule}
\title{Deletes a resolver rule}
\usage{
delete_resolver_rule(
  ResolverRuleId = "string"
)
}
\arguments{
\item{ResolverRuleId}{[required] The ID of the resolver rule that you want to delete.}
}
\description{
Deletes a resolver rule. Before you can delete a resolver rule, you must disassociate it from all the VPCs that you associated the resolver rule with. For more infomation, see DisassociateResolverRule.
}
\examples{

}
