% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{disassociate_resolver_rule}
\alias{disassociate_resolver_rule}
\title{Removes the association between a specified resolver rule and a specified VPC}
\usage{
disassociate_resolver_rule(
  VPCId = "string",
  ResolverRuleId = "string"
)
}
\arguments{
\item{VPCId}{[required] The ID of the VPC that you want to disassociate the resolver rule from.}

\item{ResolverRuleId}{[required] The ID of the resolver rule that you want to disassociate from the specified VPC.}
}
\description{
Removes the association between a specified resolver rule and a specified VPC.
}
\details{
If you disassociate a resolver rule from a VPC, Resolver stops forwarding DNS queries for the domain name that you specified in the resolver rule.
}
\examples{

}
