% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{get_resolver_rule}
\alias{get_resolver_rule}
\title{Gets information about a specified resolver rule, such as the domain name that the rule forwards DNS queries for and the ID of the outbound resolver endpoint that the rule is associated with}
\usage{
get_resolver_rule(
  ResolverRuleId = "string"
)
}
\arguments{
\item{ResolverRuleId}{[required] The ID of the resolver rule that you want to get information about.}
}
\description{
Gets information about a specified resolver rule, such as the domain name that the rule forwards DNS queries for and the ID of the outbound resolver endpoint that the rule is associated with.
}
\examples{

}
