% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{get_resolver_rule_policy}
\alias{get_resolver_rule_policy}
\title{Gets information about a resolver rule policy}
\usage{
get_resolver_rule_policy(
  Arn = "string"
)
}
\arguments{
\item{Arn}{[required] The ID of the resolver rule policy that you want to get information about.}
}
\description{
Gets information about a resolver rule policy. A resolver rule policy specifies the Resolver operations and resources that you want to allow another AWS account to be able to use.
}
\examples{

}
