% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{list_resolver_endpoint_ip_addresses}
\alias{list_resolver_endpoint_ip_addresses}
\title{Gets the IP addresses for a specified resolver endpoint}
\usage{
list_resolver_endpoint_ip_addresses(
  ResolverEndpointId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{ResolverEndpointId}{[required] The ID of the resolver endpoint that you want to get IP addresses for.}

\item{MaxResults}{The maximum number of IP addresses that you want to return in the response to a \code{ListResolverEndpointIpAddresses} request. If you don't specify a value for \code{MaxResults}, Resolver returns up to 100 IP addresses.}

\item{NextToken}{For the first \code{ListResolverEndpointIpAddresses} request, omit this value.

If the specified resolver endpoint has more than \code{MaxResults} IP addresses, you can submit another \code{ListResolverEndpointIpAddresses} request to get the next group of IP addresses. In the next request, specify the value of \code{NextToken} from the previous response.}
}
\description{
Gets the IP addresses for a specified resolver endpoint.
}
\examples{

}
