% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists the tags that you associated with the specified resource}
\usage{
list_tags_for_resource(ResourceArn, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to list tags for.}

\item{MaxResults}{The maximum number of tags that you want to return in the response to a \code{ListTagsForResource} request. If you don't specify a value for \code{MaxResults}, Resolver returns up to 100 tags.}

\item{NextToken}{For the first \code{ListTagsForResource} request, omit this value.

If you have more than \code{MaxResults} tags, you can submit another \code{ListTagsForResource} request to get the next group of tags for the resource. In the next request, specify the value of \code{NextToken} from the previous response.}
}
\description{
Lists the tags that you associated with the specified resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  ResourceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

