% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_package.R
\docType{package}
\name{paws.route53resolver-package}
\alias{paws.route53resolver}
\alias{paws.route53resolver-package}
\title{paws.route53resolver: Amazon Route 53 Resolver}
\description{
1.  Connect your network to a VPC using AWS
    Direct Connect or a VPN.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
