% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{put_resolver_rule_policy}
\alias{put_resolver_rule_policy}
\title{Specifies the Resolver operations and resources that you want to allow another AWS account to be able to use}
\usage{
put_resolver_rule_policy(
  Arn = "string",
  ResolverRulePolicy = "string"
)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the account that you want to grant permissions to.}

\item{ResolverRulePolicy}{[required] An AWS Identity and Access Management policy statement that lists the permissions that you want to grant to another AWS account.}
}
\description{
Specifies the Resolver operations and resources that you want to allow another AWS account to be able to use.
}
\examples{

}
