% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds one or more tags to a specified resource}
\usage{
tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to add tags to. To get the ARN for a resource, use the applicable \code{Get} or \code{List} command:
\itemize{
\item GetResolverEndpoint
\item GetResolverRule
\item GetResolverRuleAssociation
\item ListResolverEndpoints
\item ListResolverRuleAssociations
\item ListResolverRules
}}

\item{Tags}{[required] The tags that you want to add to the specified resource.}
}
\description{
Adds one or more tags to a specified resource.
}
\examples{

}
