% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes one or more tags from a specified resource}
\usage{
untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to remove tags from. To get the ARN for a resource, use the applicable \code{Get} or \code{List} command:
\itemize{
\item GetResolverEndpoint
\item GetResolverRule
\item GetResolverRuleAssociation
\item ListResolverEndpoints
\item ListResolverRuleAssociations
\item ListResolverRules
}}

\item{TagKeys}{[required] The tags that you want to remove to the specified resource.}
}
\description{
Removes one or more tags from a specified resource.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

