% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53resolver_operations.R
\name{update_resolver_endpoint}
\alias{update_resolver_endpoint}
\title{Updates the name of an inbound or an outbound resolver endpoint}
\usage{
update_resolver_endpoint(
  ResolverEndpointId = "string",
  Name = "string"
)
}
\arguments{
\item{ResolverEndpointId}{[required] The ID of the resolver endpoint that you want to update.}

\item{Name}{The name of the resolver endpoint that you want to update.}
}
\description{
Updates the name of an inbound or an outbound resolver endpoint.
}
\examples{

}
